build: Reduce cost of heavy compilation units
- Reduces CastExpr-tpl compilation cost by ~67%
- Parallelizes MinMaxBy registration to cut cost
and memory consumption in half
