+++ bolt/functions/prestosql/aggregates/MinByAggregateRegistration.cpp
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) ByteDance Ltd. and/or its affiliates
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "bolt/functions/prestosql/aggregates/MinMaxByAggregateRegistration.h"
+
+namespace bytedance::bolt::aggregate::prestosql {
+
+void registerMinByAggregate( // NOLINT(misc-use-internal-linkage)
+    const std::string& prefix) {
+  detail::registerMinMaxBy<
+      functions::aggregate::MinMaxByAggregateBase,
+      false,
+      MinByNAggregate>(prefix + kMinBy);
+}
+
+} // namespace bytedance::bolt::aggregate::prestosql
