+++ bolt/functions/prestosql/aggregates/MinMaxByAggregateRegistration.h
@@ -1,0 +1,108 @@
+/*
+ * Copyright (c) ByteDance Ltd. and/or its affiliates
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+#include <sstream>
+
+#include "bolt/functions/prestosql/aggregates/MinMaxByAggregates.h"
+
+namespace bytedance::bolt::aggregate::prestosql::detail {
+
+inline std::string toString(const std::vector<TypePtr>& types) {
+  std::ostringstream out;
+  for (auto i = 0; i < types.size(); ++i) {
+    if (i > 0) {
+      out << ", ";
+    }
+    out << types[i]->toString();
+  }
+  return out.str();
+}
+
+template <
+    template <
+        typename U,
+        typename V,
+        bool B1,
+        template <bool B2, typename C1, typename C2>
+        class C>
+    class Aggregate,
+    bool IsMaxFunc,
+    template <typename U, typename V>
+    class NAggregate>
+exec::AggregateRegistrationResult registerMinMaxBy(const std::string& name) {
+  std::vector<std::shared_ptr<exec::AggregateFunctionSignature>> signatures;
+  signatures.push_back(exec::AggregateFunctionSignatureBuilder()
+                           .typeVariable("V")
+                           .orderableTypeVariable("C")
+                           .returnType("V")
+                           .intermediateType("row(V,C)")
+                           .argumentType("V")
+                           .argumentType("C")
+                           .build());
+  const std::vector<std::string> supportedCompareTypes = {
+      "boolean",
+      "tinyint",
+      "smallint",
+      "integer",
+      "bigint",
+      "real",
+      "double",
+      "varchar",
+      "date",
+      "timestamp"};
+
+  for (const auto& compareType : supportedCompareTypes) {
+    signatures.push_back(exec::AggregateFunctionSignatureBuilder()
+                             .typeVariable("V")
+                             .returnType("array(V)")
+                             .intermediateType(fmt::format(
+                                 "row(bigint,array({}),array(V))", compareType))
+                             .argumentType("V")
+                             .argumentType(compareType)
+                             .argumentType("bigint")
+                             .build());
+  }
+
+  return exec::registerAggregateFunction(
+      name,
+      std::move(signatures),
+      [name](
+          core::AggregationNode::Step step,
+          const std::vector<TypePtr>& argTypes,
+          const TypePtr& resultType,
+          const core::QueryConfig&
+          /*config*/) -> std::unique_ptr<exec::Aggregate> {
+        const std::string errorMessage = fmt::format(
+            "Unknown input types for {} ({}) aggregation: {}",
+            name,
+            mapAggregationStepToName(step),
+            toString(argTypes));
+
+        const bool nAgg = (argTypes.size() == 3) ||
+            (argTypes.size() == 1 && argTypes[0]->size() == 3);
+
+        if (nAgg) {
+          return createNArg<NAggregate>(
+              resultType, argTypes[0], argTypes[1], errorMessage);
+        }
+        return functions::aggregate::create<Aggregate, Comparator, IsMaxFunc>(
+            resultType, argTypes[0], argTypes[1], errorMessage, true);
+      });
+}
+
+} // namespace bytedance::bolt::aggregate::prestosql::detail
