+++ bolt/functions/prestosql/aggregates/MinMaxByAggregates.cpp
@@ -30,102 +30,16 @@
 
 #include "bolt/functions/prestosql/aggregates/MinMaxByAggregates.h"
 namespace bytedance::bolt::aggregate::prestosql {
-std::string toString(const std::vector<TypePtr>& types) {
-  std::ostringstream out;
-  for (auto i = 0; i < types.size(); ++i) {
-    if (i > 0) {
-      out << ", ";
-    }
-    out << types[i]->toString();
-  }
-  return out.str();
-}
 
-template <
-    template <
-        typename U,
-        typename V,
-        bool B1,
-        template <bool B2, typename C1, typename C2>
-        class C>
-    class Aggregate,
-    bool isMaxFunc,
-    template <typename U, typename V>
-    class NAggregate>
-exec::AggregateRegistrationResult registerMinMaxBy(const std::string& name) {
-  std::vector<std::shared_ptr<exec::AggregateFunctionSignature>> signatures;
-  // V, C -> row(V, C) -> V.
-  signatures.push_back(exec::AggregateFunctionSignatureBuilder()
-                           .typeVariable("V")
-                           .orderableTypeVariable("C")
-                           .returnType("V")
-                           .intermediateType("row(V,C)")
-                           .argumentType("V")
-                           .argumentType("C")
-                           .build());
-  const std::vector<std::string> supportedCompareTypes = {
-      "boolean",
-      "tinyint",
-      "smallint",
-      "integer",
-      "bigint",
-      "real",
-      "double",
-      "varchar",
-      "date",
-      "timestamp"};
+void registerMaxByAggregate( // NOLINT(misc-use-internal-linkage)
+    const std::string& prefix);
+void registerMinByAggregate( // NOLINT(misc-use-internal-linkage)
+    const std::string& prefix);
 
-  // Add support for all value types to 3-arg version of the aggregate.
-  for (const auto& compareType : supportedCompareTypes) {
-    // V, C, bigint -> row(bigint, array(C), array(V)) -> array(V).
-    signatures.push_back(exec::AggregateFunctionSignatureBuilder()
-                             .typeVariable("V")
-                             .returnType("array(V)")
-                             .intermediateType(fmt::format(
-                                 "row(bigint,array({}),array(V))", compareType))
-                             .argumentType("V")
-                             .argumentType(compareType)
-                             .argumentType("bigint")
-                             .build());
-  }
-
-  return exec::registerAggregateFunction(
-      name,
-      std::move(signatures),
-      [name](
-          core::AggregationNode::Step step,
-          const std::vector<TypePtr>& argTypes,
-          const TypePtr& resultType,
-          const core::QueryConfig&
-          /*config*/) -> std::unique_ptr<exec::Aggregate> {
-        const std::string errorMessage = fmt::format(
-            "Unknown input types for {} ({}) aggregation: {}",
-            name,
-            mapAggregationStepToName(step),
-            toString(argTypes));
-
-        const bool nAgg = (argTypes.size() == 3) ||
-            (argTypes.size() == 1 && argTypes[0]->size() == 3);
-
-        if (nAgg) {
-          return createNArg<NAggregate>(
-              resultType, argTypes[0], argTypes[1], errorMessage);
-        } else {
-          return functions::aggregate::create<Aggregate, Comparator, isMaxFunc>(
-              resultType, argTypes[0], argTypes[1], errorMessage, true);
-        }
-      });
-}
-
-void registerMinMaxByAggregates(const std::string& prefix) {
-  registerMinMaxBy<
-      functions::aggregate::MinMaxByAggregateBase,
-      true,
-      MaxByNAggregate>(prefix + kMaxBy);
-  registerMinMaxBy<
-      functions::aggregate::MinMaxByAggregateBase,
-      false,
-      MinByNAggregate>(prefix + kMinBy);
+void registerMinMaxByAggregates( // NOLINT(misc-use-internal-linkage)
+    const std::string& prefix) {
+  registerMaxByAggregate(prefix);
+  registerMinByAggregate(prefix);
 }
 
 } // namespace bytedance::bolt::aggregate::prestosql
